refactored mint fiat
