+++ pages/create-position/[collateralId]/open/index.tsx
@@ -16,10 +16,7 @@ import { CreatePositionUnderlying } from '@/src/components/custom/create-positio
 import { CreatePositionBond } from '@/src/components/custom/create-position-bond'
 import { PositionFormsLayout } from '@/src/components/custom/position-forms-layout'
 import { Summary } from '@/src/components/custom/summary'
-import {
-  VIRTUAL_RATE_MAX_SLIPPAGE,
-  WAD_DECIMALS,
-} from '@/src/constants/misc'
+import { WAD_DECIMALS } from '@/src/constants/misc'
 import { useDynamicTitle } from '@/src/hooks/useDynamicTitle'
 import { useERC20Allowance } from '@/src/hooks/useERC20Allowance'
 import useUserProxy from '@/src/hooks/useUserProxy'
@@ -30,7 +27,7 @@ import { useQueryParam } from '@/src/hooks/useQueryParam'
 import { useCollateral } from '@/src/hooks/subgraph/useCollateral'
 import { Collateral } from '@/src/utils/data/collaterals'
 import { parseDate } from '@/src/utils/dateTime'
-import { ONE_BIG_NUMBER, ZERO_BIG_NUMBER } from '@/src/constants/misc'
+import { ZERO_BIG_NUMBER } from '@/src/constants/misc'
 import { getHumanValue, getNonHumanValue, perSecondToAPR } from '@/src/web3/utils'
 import { useTokenDecimalsAndBalance } from '@/src/hooks/useTokenDecimalsAndBalance'
 import SuccessAnimation from '@/src/resources/animations/success-animation.json'
@@ -100,25 +97,6 @@ const FormERC20: React.FC<{
   const [activeMachine, setActiveMachine] = useState(stateMachine)
 
   const [tab, setTab] = useState('bond')
-
-  // @TODO: not working max amount
-  // maxFIAT = totalCollateral*collateralValue/collateralizationRatio/(virtualRateSafetyMargin*virtualRate)-debt
-  const maxBorrowAmountCalculated = useMemo(() => {
-    const totalCollateral = stateMachine.context.erc20Amount ?? ZERO_BIG_NUMBER
-    const collateralValue = getHumanValue(collateral.currentValue || ONE_BIG_NUMBER, WAD_DECIMALS)
-    const collateralizationRatio = getHumanValue(
-      collateral.vault.collateralizationRatio || ONE_BIG_NUMBER,
-      WAD_DECIMALS,
-    )
-
-    const virtualRateWithMargin = VIRTUAL_RATE_MAX_SLIPPAGE.times(collateral.vault.virtualRate)
-    const maxBorrowAmount = totalCollateral
-      .times(collateralValue)
-      .div(collateralizationRatio)
-      .div(virtualRateWithMargin)
-
-    return maxBorrowAmount
-  }, [stateMachine.context.erc20Amount, collateral])
 
   // @TODO: ui should show that the minimum fiat to have in a position is the debtFloor
   const hasMinimumFIAT = useMemo(() => {
@@ -201,7 +179,6 @@ const FormERC20: React.FC<{
                 collateral={collateral}
                 loading={loading}
                 healthFactorNumber={healthFactorNumber}
-                maxBorrowAmountCalculated={maxBorrowAmountCalculated}
                 hasMinimumFIAT={hasMinimumFIAT}
                 setLoading={setFormLoading}
                 setMachine={switchActiveMachine}
@@ -211,7 +188,6 @@ const FormERC20: React.FC<{
               <CreatePositionBond
                 collateral={collateral}
                 loading={loading}
-                maxBorrowAmountCalculated={maxBorrowAmountCalculated}
                 isDisabledCreatePosition={isDisabledCreatePosition}
                 setLoading={setFormLoading}
                 setMachine={switchActiveMachine}
