+++ src/components/custom/mint-fiat/s.module.scss
@@ -1,0 +1,3 @@
+.mintFiat {
+    margin-bottom: 26px;
+}
